+++ cpp/src/train/vgg16/train_vgg16.h
@@ -3,11 +3,11 @@
 #include "train/train_model.h"
 
 // Minimum size required by VGG-16 for the feature vector not to cancel out
-constexpr int64_t kVggMinImageSize = 32;
+constexpr int32_t kVggMinImageSize = 32;
 
 template <typename Dataset>
 void train_vgg16(const char* dataRootRelativePath, const char* classesJson, const char* vgg_dataset_filename,
-    const int64_t trainBatchSize, const int64_t testBatchSize, const int64_t numberOfEpochs)
+    const int32_t trainBatchSize, const int32_t testBatchSize, const int32_t numberOfEpochs)
 {
     train_model<Dataset>(dataRootRelativePath, classesJson,
         vgg_dataset_filename, kVggMinImageSize,
